fix: prelude in starting card selection
