+++ backend/cmd/server/main.go
@@ -136,9 +136,7 @@ func main() {
 	// Turn management (5)
 	startGameAction := turnAction.NewStartGameAction(gameRepo, log)
 	skipActionAction := turnAction.NewSkipActionAction(gameRepo, finalScoringAction, log)
-	selectCorporationAction := turnAction.NewSelectCorporationAction(gameRepo, cardRegistry, log)
-	selectStartingCardsAction := turnAction.NewSelectStartingCardsAction(gameRepo, cardRegistry, log)
-	selectPreludeCardsAction := turnAction.NewSelectPreludeCardsAction(gameRepo, cardRegistry, log)
+	selectStartingChoicesAction := turnAction.NewSelectStartingChoicesAction(gameRepo, cardRegistry, log)
 
 	// Confirmations (4)
 	confirmSellPatentsAction := confirmAction.NewConfirmSellPatentsAction(gameRepo, log)
@@ -198,9 +196,7 @@ func main() {
 		// Turn management
 		startGameAction,
 		skipActionAction,
-		selectCorporationAction,
-		selectStartingCardsAction,
-		selectPreludeCardsAction,
+		selectStartingChoicesAction,
 		// Confirmations
 		confirmSellPatentsAction,
 		confirmProductionCardsAction,
