+++ backend/internal/delivery/dto/game_dto.go
@@ -5,11 +5,9 @@ type GamePhase string
 
 const (
 	GamePhaseWaitingForGameStart   GamePhase = "waiting_for_game_start"
-	GamePhaseCorporationSelection  GamePhase = "corporation_selection"
-	GamePhaseStartingCardSelection GamePhase = "starting_card_selection"
+	GamePhaseStartingSelection     GamePhase = "starting_selection"
 	GamePhaseStartGameSelection    GamePhase = "start_game_selection"
 	GamePhaseDemoSetup             GamePhase = "demo_setup"
-	GamePhasePreludeSelection      GamePhase = "prelude_selection"
 	GamePhaseAction                GamePhase = "action"
 	GamePhaseProductionAndCardDraw GamePhase = "production_and_card_draw"
 	GamePhaseComplete              GamePhase = "complete"
