+++ frontend/src/components/ui/overlay/StartingCardSelectionOverlay.tsx
@@ -1,7 +1,9 @@
-import React from "react";
+import React, { useState, useCallback, useEffect, useMemo } from "react";
+import CorporationCard from "../cards/CorporationCard.tsx";
 import GameCard from "../cards/GameCard.tsx";
 import GameIcon from "../display/GameIcon.tsx";
 import { CardDto, ResourceTypeCredit } from "../../../types/generated/api-types.ts";
+import { getCorporationBorderColor } from "@/utils/corporationColors.ts";
 import { useCardSelection } from "../../../hooks/useCardSelection.ts";
 import {
   OVERLAY_CONTAINER_CLASS,
@@ -13,69 +15,197 @@ import {
   RESOURCE_DISPLAY_CLASS,
 } from "./overlayStyles.ts";
 import GameMenuButton from "../buttons/GameMenuButton.tsx";
+import MainMenuSettingsButton from "../buttons/MainMenuSettingsButton.tsx";
 
 interface StartingCardSelectionOverlayProps {
   isOpen: boolean;
+  availableCorporations: CardDto[];
+  availablePreludes: CardDto[];
+  maxSelectablePreludes: number;
   cards: CardDto[];
   playerCredits: number;
-  onSelectCards: (selectedCardIds: string[]) => void;
+  onConfirm: (corporationId: string, preludeIds: string[], cardIds: string[]) => void;
 }
 
 const StartingCardSelectionOverlay: React.FC<StartingCardSelectionOverlayProps> = ({
   isOpen,
+  availableCorporations,
+  availablePreludes,
+  maxSelectablePreludes,
   cards,
   playerCredits,
-  onSelectCards,
+  onConfirm,
 }) => {
+  const [selectedCorporationId, setSelectedCorporationId] = useState<string | null>(null);
+  const [selectedPreludeIds, setSelectedPreludeIds] = useState<string[]>([]);
+
+  const selectedCorp = useMemo(
+    () => availableCorporations.find((c) => c.id === selectedCorporationId),
+    [availableCorporations, selectedCorporationId],
+  );
+
+  const effectiveCredits = selectedCorp?.startingResources?.credits ?? playerCredits;
+
   const {
     selectedCardIds,
     totalCost,
     showConfirmation,
-    isValidSelection,
+    isValidSelection: isValidCardSelection,
     handleCardSelect,
     handleConfirm: handleCardConfirm,
   } = useCardSelection({
     cards,
     isOpen,
-    playerCredits,
+    playerCredits: effectiveCredits,
     costPerCard: 3,
     minCards: 0,
   });
 
-  if (!isOpen || cards.length === 0) return null;
+  useEffect(() => {
+    if (isOpen) {
+      setSelectedCorporationId(null);
+      setSelectedPreludeIds([]);
+    }
+  }, [isOpen]);
+
+  const handlePreludeSelect = useCallback(
+    (cardId: string) => {
+      setSelectedPreludeIds((prev) => {
+        if (prev.includes(cardId)) {
+          return prev.filter((id) => id !== cardId);
+        }
+        if (prev.length >= maxSelectablePreludes) {
+          return prev;
+        }
+        return [...prev, cardId];
+      });
+    },
+    [maxSelectablePreludes],
+  );
+
+  const hasPreludes = availablePreludes.length > 0;
+  const preludesValid = !hasPreludes || selectedPreludeIds.length === maxSelectablePreludes;
+  const allValid = !!selectedCorporationId && preludesValid && isValidCardSelection;
 
   const handleConfirm = () => {
+    if (!selectedCorporationId) return;
+
     handleCardConfirm((cardIds) => {
-      onSelectCards(cardIds);
+      onConfirm(selectedCorporationId, selectedPreludeIds, cardIds);
     });
   };
 
+  if (!isOpen) return null;
+
   return (
     <div className="fixed inset-0 z-[1000] flex items-center justify-center animate-[fadeIn_0.3s_ease]">
+      <MainMenuSettingsButton />
       <div className={OVERLAY_CONTAINER_CLASS}>
         <div className={OVERLAY_HEADER_CLASS}>
           <h2 className={OVERLAY_TITLE_CLASS}>Select Starting Cards</h2>
           <p className={OVERLAY_DESCRIPTION_CLASS}>
-            Choose your starting cards. Each card costs 3 MC.
+            Choose your corporation
+            {hasPreludes ? ", prelude cards," : ""} and starting project cards. Each project card
+            costs 3 MC.
           </p>
         </div>
 
-        <div className="flex-1 overflow-x-auto overflow-y-hidden p-8 flex items-center bg-[radial-gradient(ellipse_at_center,rgba(139,69,19,0.1)_0%,transparent_70%)] [&::-webkit-scrollbar]:h-2 [&::-webkit-scrollbar-track]:bg-white/5 [&::-webkit-scrollbar-track]:rounded [&::-webkit-scrollbar-thumb]:bg-white/20 [&::-webkit-scrollbar-thumb]:rounded [&::-webkit-scrollbar-thumb:hover]:bg-white/30 max-[768px]:p-5">
-          <div className="flex gap-6 mx-auto py-5 max-[768px]:gap-4">
-            {cards.map((card, index) => {
-              const isSelected = selectedCardIds.includes(card.id);
+        <div className="flex-1 overflow-y-auto p-8 bg-black/20 max-[768px]:p-5">
+          {availableCorporations.length > 0 && (
+            <div>
+              <h3 className="text-white/60 text-sm font-orbitron font-bold uppercase tracking-widest mb-4">
+                Corporation
+              </h3>
+              <div className="flex gap-4 justify-center flex-wrap">
+                {availableCorporations.map((corp) => (
+                  <div key={corp.id} className="w-[400px] max-[768px]:w-full">
+                    <CorporationCard
+                      corporation={{
+                        id: corp.id,
+                        name: corp.name,
+                        description: corp.description,
+                        startingMegaCredits: corp.startingResources?.credits ?? 0,
+                        startingProduction: corp.startingProduction
+                          ? {
+                              credits: corp.startingProduction.credits,
+                              steel: corp.startingProduction.steel,
+                              titanium: corp.startingProduction.titanium,
+                              plants: corp.startingProduction.plants,
+                              energy: corp.startingProduction.energy,
+                              heat: corp.startingProduction.heat,
+                            }
+                          : undefined,
+                        startingResources: corp.startingResources
+                          ? {
+                              credits: corp.startingResources.credits,
+                              steel: corp.startingResources.steel,
+                              titanium: corp.startingResources.titanium,
+                              plants: corp.startingResources.plants,
+                              energy: corp.startingResources.energy,
+                              heat: corp.startingResources.heat,
+                            }
+                          : undefined,
+                        behaviors: corp.behaviors,
+                        logoPath: undefined,
+                      }}
+                      isSelected={selectedCorporationId === corp.id}
+                      onSelect={setSelectedCorporationId}
+                      borderColor={getCorporationBorderColor(corp.name)}
+                    />
+                  </div>
+                ))}
+              </div>
+            </div>
+          )}
 
-              return (
-                <GameCard
-                  key={card.id}
-                  card={card}
-                  isSelected={isSelected}
-                  onSelect={handleCardSelect}
-                  animationDelay={index * 100}
-                  showCheckbox={true}
-                />
-              );
-            })}
+          {hasPreludes && (
+            <>
+              <div className="border-t border-white/10 my-6" />
+              <div>
+                <h3 className="text-white/60 text-sm font-orbitron font-bold uppercase tracking-widest mb-4">
+                  Prelude Cards
+                  <span className="ml-2 text-white/40 text-xs normal-case">
+                    {selectedPreludeIds.length} / {maxSelectablePreludes} selected
+                  </span>
+                </h3>
+                <div className="flex gap-6 justify-center flex-wrap max-[768px]:gap-4">
+                  {availablePreludes.map((card, index) => (
+                    <GameCard
+                      key={card.id}
+                      card={card}
+                      isSelected={selectedPreludeIds.includes(card.id)}
+                      onSelect={handlePreludeSelect}
+                      animationDelay={index * 100}
+                      showCheckbox={true}
+                    />
+                  ))}
+                </div>
+              </div>
+            </>
+          )}
+
+          <div className="border-t border-white/10 my-6" />
+
+          <div>
+            <h3 className="text-white/60 text-sm font-orbitron font-bold uppercase tracking-widest mb-4">
+              Project Cards
+            </h3>
+            <div className="flex gap-6 justify-center flex-wrap max-[768px]:gap-4">
+              {cards.map((card, index) => {
+                const isSelected = selectedCardIds.includes(card.id);
+
+                return (
+                  <GameCard
+                    key={card.id}
+                    card={card}
+                    isSelected={isSelected}
+                    onSelect={handleCardSelect}
+                    animationDelay={index * 100}
+                    showCheckbox={true}
+                  />
+                );
+              })}
+            </div>
           </div>
         </div>
 
@@ -83,7 +213,7 @@ const StartingCardSelectionOverlay: React.FC<StartingCardSelectionOverlayProps> 
           <div className="flex gap-8 items-center max-[768px]:w-full max-[768px]:justify-between">
             <div className={RESOURCE_DISPLAY_CLASS}>
               <span className={RESOURCE_LABEL_CLASS}>Your Credits:</span>
-              <GameIcon iconType={ResourceTypeCredit} amount={playerCredits} size="large" />
+              <GameIcon iconType={ResourceTypeCredit} amount={effectiveCredits} size="large" />
             </div>
             <div className={RESOURCE_DISPLAY_CLASS}>
               <span className={RESOURCE_LABEL_CLASS}>Total Cost:</span>
@@ -102,6 +232,10 @@ const StartingCardSelectionOverlay: React.FC<StartingCardSelectionOverlayProps> 
           </div>
 
           <div className="flex items-center gap-4 max-[768px]:w-full max-[768px]:flex-col max-[768px]:gap-3">
+            {!selectedCorporationId && (
+              <span className="text-sm text-white/70">Select a corporation to continue</span>
+            )}
+
             {showConfirmation && (
               <div className="text-sm">
                 <span className="text-[#ff9800]">
@@ -114,7 +248,7 @@ const StartingCardSelectionOverlay: React.FC<StartingCardSelectionOverlayProps> 
               variant="primary"
               size="lg"
               onClick={handleConfirm}
-              disabled={!isValidSelection}
+              disabled={!allValid}
               className="whitespace-nowrap max-[768px]:w-full max-[768px]:py-3 max-[768px]:px-6 max-[768px]:text-lg"
             >
               {showConfirmation ? "Confirm Skip" : "Confirm Selection"}
