+++ infra/.env.example
@@ -4,7 +4,7 @@ TM_LOG_LEVEL=info
 # Bug reporting -- GitHub App (optional)
 GITHUB_APP_ID=
 GITHUB_INSTALLATION_ID=
-GITHUB_PRIVATE_KEY_FILE=/path/to/your/private-key.pem
+GITHUB_PRIVATE_KEY_FILE=
 GITHUB_REPO_OWNER=
 GITHUB_REPO_NAME=
